refactor(measures): rename tier_0/ → tier_0_pre_commit/; runner accepts both forms
Tier_0 is the only tier with two practical invocation contexts: the
existing capture-ci scheduled runtime (pre-commit + test:local) and the
forthcoming agent-hook PostToolUse runtime (single-file, sub-100ms).
Renaming the scheduled half to tier_0_pre_commit/ frees tier_0_post_edit/
to land alongside without conflating contracts.

Discovery in capture-ci-checks.ts now scans tier_N AND tier_N_pre_commit
(both forms accepted; bare tier_N stays the convention for tiers 1-4 which
only have one invocation context). CHECK_PATH regex tolerates the suffix
so tierFromMeasurePath / concernFromMeasurePath keep working.

Behavior-preserving: same checks, same schedule, same exit codes.
