+++ packages/measures/src/checks/tier_0_pre_commit/static/agent-instructions-sync.mjs
