+++ packages/measures/src/checks/tier_0_pre_commit/static/agent-instructions-sync.ts
@@ -1,6 +1,6 @@
 import {type CheckDef} from '../../../_types.ts'
 
-const SCRIPT_PATH = 'packages/measures/src/checks/tier_0/static/agent-instructions-sync.mjs'
+const SCRIPT_PATH = 'packages/measures/src/checks/tier_0_pre_commit/static/agent-instructions-sync.mjs'
 
 export const check: CheckDef = {
     id: 'agent-instructions-sync',
